[doc] Update the README.md file
This is a backport from elemental-expath-package-parent (https://github.com/evolvedbinary/elemental-expath-package-parent/commit/0a831f5811e416fefeda8046a917fe4b9f2268c8) under the LGPL 2.1 only License. Only the ported code as it appears in this commit, at the time that it was contributed to the eXist-db project, was re-licensed under The LGPL 2.1 only License by Evolved Binary for use in exist-apps-parent.